+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Show push time in human-readable format in the current locale's timezone
 - Display an informative error message when attempting to push an empty base repo
 - Allow users to assign multiple branches in one call to `assign`
+- Print help for the subcommand when `assigner help <command>` is run
 
 ## 0.1.0